dep: remove shaderc dependency
