+++ librashader-reflect/Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["shader", "retroarch", "SPIR-V"]
 description = "RetroArch shaders for all."
 
 [dependencies]
-shaderc = { version = "0.8.3", features = [] }
+glslang = "0.2"
 bytemuck = "1.13.0"
 
 thiserror = "1.0.37"
@@ -40,7 +40,6 @@ optional = true
 
 [features]
 default = ["cross", "wgsl", "serialize"]
-standalone = ["shaderc/build-from-source", "shaderc/prefer-static-linking"]
 dxil = ["cross", "spirv-to-dxil"]
 wgsl = ["cross", "naga", "spirv", "rspirv"]
 cross = [ "spirv_cross", "spirv_cross/glsl", "spirv_cross/hlsl" ]
