+++ librashader-runtime-d3d11/Cargo.toml
@@ -15,7 +15,7 @@ description = "RetroArch shaders for all."
 librashader-common = { path = "../librashader-common", features = ["d3d11"], version = "0.2.0-beta.6" }
 librashader-presets = { path = "../librashader-presets", version = "0.2.0-beta.6" }
 librashader-preprocess = { path = "../librashader-preprocess", version = "0.2.0-beta.6" }
-librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6", features = ["standalone"]  }
+librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6"}
 librashader-runtime = { path = "../librashader-runtime", version = "0.2.0-beta.6" }
 spirv_cross = { package = "librashader-spirv-cross", version = "0.23" }
 librashader-cache = { path = "../librashader-cache", version = "0.2.0-beta.6", features = ["d3d"] }
