+++ librashader-runtime-vk/Cargo.toml
@@ -17,7 +17,7 @@ description = "RetroArch shaders for all."
 librashader-common = { path = "../librashader-common", features = ["vulkan"], version = "0.2.0-beta.6" }
 librashader-presets = { path = "../librashader-presets", version = "0.2.0-beta.6" }
 librashader-preprocess = { path = "../librashader-preprocess", version = "0.2.0-beta.6" }
-librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6", features = []   }
+librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6"  }
 librashader-runtime = { path = "../librashader-runtime" , version = "0.2.0-beta.6" }
 librashader-cache = { path = "../librashader-cache", version = "0.2.0-beta.6" }
