+++ librashader/Cargo.toml
@@ -16,7 +16,7 @@ description = "RetroArch shaders for all."
 librashader-common = { path = "../librashader-common", version = "0.2.0-beta.6" }
 librashader-presets = { path = "../librashader-presets", version = "0.2.0-beta.6" }
 librashader-preprocess = { path = "../librashader-preprocess", version = "0.2.0-beta.6" }
-librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6", features = ["standalone"] }
+librashader-reflect = { path = "../librashader-reflect", version = "0.2.0-beta.6" }
 librashader-runtime  = { path = "../librashader-runtime", version = "0.2.0-beta.6" }
 librashader-runtime-d3d11  = { path = "../librashader-runtime-d3d11", version = "0.2.0-beta.6", optional = true }
 librashader-runtime-d3d12  = { path = "../librashader-runtime-d3d12", version = "0.2.0-beta.6", optional = true }
