+++ pkg/librashader.spec
@@ -12,7 +12,6 @@ URL:      https://github.com/SnowflakePowered/%{name}
 %undefine _disable_source_fetch
 Source:  https://github.com/SnowflakePowered/%{name}/archive/%{commit}/%{name}-%{shortcommit}.tar.gz
 BuildRequires: pkgconfig(vulkan)
-BuildRequires: pkgconfig(shaderc)
 BuildRequires: cmake
 BuildRequires: gcc
 BuildRequires: g++
