+++ pkg/obs/librashader.spec
@@ -10,7 +10,6 @@ Source0:  librashader-%{version}.tar.xz
 Source1:  vendor.tar.xz
 Source2:  cargo_config
 BuildRequires: pkgconfig(vulkan)
-BuildRequires: pkgconfig(shaderc)
 BuildRequires: ninja-build
 BuildRequires: patchelf
 BuildRequires: cmake
